+++ src/Internal/Container/ContainerInterface.php
@@ -29,5 +29,7 @@ interface ContainerInterface
 
     public function move(string $source, string $destination): void;
 
+    public function prepareForSourceReplacement(): void;
+
     public function saveAs(string $filename): void;
 }